requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.1
opencv-python==4.8.1.78
numpy==1.26.1
scipy==1.11.1
pandas==2.1.1
dlib==19.24.1
mediapipe==0.10.12
python-multipart==0.0.6       
requests==2.34.0 
langchain
langchain-community
langchain-core
langchain-groq
langchain-huggingface
python-dotenv
faiss-cpu
sentence-transformers
torch